% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/docs.R
\name{.format_docs_summary_md}
\alias{.format_docs_summary_md}
\title{Format Docs Summary as Markdown}
\usage{
.format_docs_summary_md(package)
}
\arguments{
\item{package}{Character. Package name.}
}
\value{
Character string of markdown
}
\description{
Creates a compact summary of package documentation.
}
\keyword{internal}
