% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/fyi.R
\name{.format_exports_md}
\alias{.format_exports_md}
\title{Format Exports as Markdown}
\usage{
.format_exports_md(df, package)
}
\arguments{
\item{df}{Data.frame from fyi_exports()}

\item{package}{Package name for header}
}
\value{
Character string of markdown
}
\description{
Format Exports as Markdown
}
\keyword{internal}
