% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/options.R
\name{.format_options_md}
\alias{.format_options_md}
\title{Format Options as Markdown}
\usage{
.format_options_md(df, package)
}
\arguments{
\item{df}{Data.frame from fyi_options()}

\item{package}{Package name for header}
}
\value{
Character string of markdown
}
\description{
Format Options as Markdown
}
\keyword{internal}
