% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/rd2md.R
\name{.rd_arguments_to_md}
\alias{.rd_arguments_to_md}
\title{Convert \\\\arguments Section to Markdown}
\usage{
.rd_arguments_to_md(element)
}
\description{
Convert \\\\arguments Section to Markdown
}
\keyword{internal}
