% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/rd2md.R
\name{.rd_content_to_md}
\alias{.rd_content_to_md}
\title{Convert Rd Content to Markdown}
\usage{
.rd_content_to_md(element)
}
\description{
Handles Rd markup like \\code{}, \\link{}, \\emph{}, etc.
}
\keyword{internal}
