\name{fyi_help_topics}
\alias{fyi_help_topics}
\title{List Help Topics for a Package}
\description{Returns all documented topics (help pages) in a package.}
\usage{
fyi_help_topics(package)
}
\arguments{
  \item{package}{Character. Package name.}
}
\value{
A character vector of help topic names.
}
\examples{
\dontrun{
fyi_help_topics("sttapi")
}
}
