\name{.format_docs_summary_md}
\alias{.format_docs_summary_md}
\title{Format Docs Summary as Markdown}
\description{Creates a compact summary of package documentation.}
\usage{
.format_docs_summary_md(package)
}
\arguments{
  \item{package}{Character. Package name.}
}
\value{
Character string of markdown
}
\keyword{internal}
