\name{.format_internals_md}
\alias{.format_internals_md}
\title{Format Internals as Markdown}
\description{Format Internals as Markdown}
\usage{
.format_internals_md(df, package)
}
\arguments{
  \item{df}{Data.frame from fyi_internals()}
  \item{package}{Package name for header}
}
\value{
Character string of markdown
}
\keyword{internal}
