% tinyrox says don't edit this manually, but it can't stop you!
\name{fyi_help_topics}
\alias{fyi_help_topics}
\title{List Help Topics for a Package}
\usage{
fyi_help_topics(package)
}
\arguments{
\item{package}{Character. Package name.}
}
\value{
A character vector of help topic names.
}
\description{
Returns all documented topics (help pages) in a package.
}
\examples{
\dontrun{
fyi_help_topics("sttapi")
}
}
